# openapi/src/openapi.yaml
openapi: 3.0.3
info:
  title: 南大软件测试中心在线系统
  description: 后端 API 文档
  version: 1.0.0
  contact:
    name: 在线系统项目 A 组
    url: https://github.com/OnlineSystemGroupA
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://masterchedan.top:8081/
    description: Development server
paths:
  /auth/register:
    post:
      tags:
        - authentication
      summary: 用户注册
      description: 注册新用户
      operationId: register
      requestBody:
        description: "用户注册所必需的信息"
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/RegisterParam"
        required: true
      responses:

        '200':
          description: 注册成功
        '409':
          description: 用户名已存在
  /auth/login:
    post:
      tags:
        - authentication
      summary: 用户登录
      description: 登录成功返回 token
      operationId: login
      parameters:
        - name: user-type
          in: query
          required: true
          description: 进行登录的用户类型：admin、operator、client.
          schema:
            type: string
      requestBody:
        description: 用户名、密码
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/LoginParam"
        required: true
      responses:
        '200':
          description: 登录成功
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Token"
        '404':
          description: 用户不存在
        '401':
          description: 用户名或密码错误
        '403':
          description: 账号禁用
...
  /archive/processes:
    get:
      tags:
        - archive
      summary: 获取已完成流程实例
      description: 获取与当前用户相关的所有已完成流程实例
      operationId: getArchivedProcesses
      parameters:
        - name: assigned
          in: query
          required: false
          description: 其值为 true 表示获取当前有任务被分配给当前用户的项目
          schema:
            type: boolean
        - name: pageIndex
          in: query
          required: true
          description: 需要查询的页号
          schema:
            type: integer
            format: int32
        - name: numPerPage
          in: query
          required: true
          description: 每页的项目条目数
          schema:
            type: integer
            format: int32
        - name: orderBy
          in: query
          required: true
          description: 用于排序的字段
          schema:
            type: string
      responses:
        '200':
          description: 成功获取流程实例列表
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Process"
  /archive/processes/{projectId}/details:
    parameters:
      - name: projectId
        in: path
        description: 流程实例 Id
        required: true
        schema:
          type: integer
          format: int64
    delete:
      tags:
        - admin
      summary: 删除已完成流程
      description: 管理员删除已完成流程
      operationId: deleteArchiveProcess
      responses:
        '200':
          description: 成功删除指定流程
        '403':
          description: 指定流程对该用户不可见
        '404':
          description: 指定流程不存在
    get:
      summary: 获取流程记录
      description: 获取流程记录
      operationId: getProcessRecord
      responses:
        '200':
          description: ok
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/ProcessDetails"
        '403':
          description: 指定流程对该用户不可见
        '404':
          description: 指定流程不存在
...
  /archive/processes/{projectId}/forms:
    get:
      tags:
        - archive
      summary: 获取表单
      description: 获取表单
      operationId: getArchivedFormInfo
      parameters:
        - name: projectId
          in: path
          description: 指定流程记录的项目号
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: ok
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormInfo"
        '404':
          description: 指定流程记录不存在
  /archive/processes/{projectId}/forms/{formName}:
    parameters:
      - name: projectId
        in: path
        description: 指定流程实例 id
        schema:
          type: integer
          format: int64
        required: true
      - name: formName
        in: path
        description: 期望操作的表单名称
        schema:
          type: string
        required: true
    get:
      tags:
        - archive
      summary: 获取表单
      description: 获取表单
      operationId: getArchivedForm
      responses:
        '200':
          description: 成功获取指定表单
          content:
            'text/plain':
              schema:
                description: 包含表单对象的 JSON 字符串
                type: string
        '403':
          description: 指定流程或表单对该用户不可见
        '404':
          description: 指定流程或表单不存在
...
components:
  schemas:
...
    RegisterParam:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          description: 用户名，唯一
        email:
          type: string
          description: 用户邮箱
        password:
          type: string
          description: 用户密码
    LoginParam:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    Token:
      type: object
      required:
        - tokenHead
        - tokenStr
      properties:
        tokenHead:
          type: string
          description: token 的头部，前端请求携带 token 时需要将其拼接在 token 之前
        tokenStr:
          type: string
          description: token 字符串
    Process:
      type: object
      required:
        - projectId
        - processId
        - taskId
        - title
        - taskName
        - assignee
        - startUser
        - startDate
        - endDate
      properties:
        projectId:
          type: string
          description: 软件测试项目的 ID
        processId:
          type: string
          description: 流程实例 ID
        taskId:
          type: string
          description: 当前任务的实例 ID
        title:
          type: string
          description: 待测试软件名称
        taskName:
          type: string
          description: 当前任务的名称
        assignee:
          type: string
          description: 当前任务被分配人姓名
        startUser:
          type: string
          description: 流程发起人姓名
        startDate:
          type: string
          description: 流程发起时间
        endDate:
          type: string
          description: 流程结束时间
...
    ProcessDetails:
      required:
        - projectId
        - title
        - version
        - testType
        - applicationDate
        - applicant
        - company
        - telephone
        - email
        - address
        - startDate
        - dueDate
        - index
        - currentTaskName
        - assignee
      properties:
        projectId:
          description: 测试项目号
          type: integer
          format: int64
        title:
          description: 软件项目名
          type: string
        version:
          description: 软件版本
          type: string
        testType:
          description: 测试类型
          type: string
        applicationDate:
          description: 申请日期
          type: string
        applicant:
          description: 申请人
          type: string
        company:
          description: 公司名称
          type: string
        telephone:
          description: 电话
          type: string
        email:
          description: 邮箱
          type: string
        address:
          description: 地址
          type: string
        startDate:
          description: 测试开始时间
          type: string
        dueDate:
          description: 预计结束时间
          type: string
        index:
          description: 当前任务所在组的编号
          type: integer
          format: int32
        currentTaskName:
          description: 当前任务名称
          type: string
        assignee:
          description: 当前被分配人姓名
          type: string
...
    FormInfo:
      required:
        - formType
        - createDate
        - formState
      properties:
        formType:
          type: string
          description: 表单类型
        createDate:
          type: string
          description: 表单创建日期
        formState:
          type: string
          description: 表单当前状态

    # 账号相关 DTO
...